replaced premake with cmake, replaced collision helper functions with class that exclusively manages colliisons
