+++ .gitignore
@@ -2,5 +2,6 @@
 .vscode/*
 
 */bin/*
+build
 Makefile
 *.make
