+++ physics-engine/include/collision/collider-base.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "math/basic-data-types.h"
+
+namespace physics {
+
+class ColliderBase
+{
+public:
+    virtual inline vector3f furthestPoint(const vector3f& dir) = 0;
+};
+
+} // namespace physics
